+++ src/main/java/com/bugtracker/alpha/entities/Company.java
@@ -7,22 +7,22 @@ import javax.persistence.*;
 public class Company {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int company_id;
+  private long company_id;
   private String name;
 
   public Company() {
 
   }
-
+  
   public Company(String name) {
     this.name = name;
   }
   
-  public int getCompany_id() {
+  public long getCompany_id() {
     return this.company_id;
   }
 
-  public void setCompany_id(int company_id) {
+  public void setCompany_id(long company_id) {
     this.company_id = company_id;
   }
